move App to app/
