+++ src/index.tsx
@@ -1,7 +1,7 @@
 import { CssVarsProvider, CssBaseline, extendTheme } from '@mui/joy'
 import React from 'react'
 import ReactDOM from 'react-dom/client'
-import App from './App'
+import App from './app/App'
 import { ApolloClient, InMemoryCache, ApolloProvider } from '@apollo/client'
 import tauriGraphqlApolloLink from './tauri/tauriGraphqlApolloLink'
 // import tauriGraphqlHttpLink from './tauri/tauriGraphqlHttpLink'
